Add menu to navbar
