+++ src/components/RestaurantCard/RestaurantCard.css
@@ -1,0 +1,52 @@
+.product-item {
+  background-color: #25274d;
+  border: none;
+  border-radius: 10px;
+  overflow: hidden;
+}
+
+.product-image {
+  width: 100%;
+  height: 200px;
+  object-fit: cover;
+}
+
+.product-info {
+  padding: 10px;
+  text-align: center;
+}
+
+.product-name {
+  font-size: 1.2rem;
+  color: #fff;
+  margin: 0;
+}
+
+.product-pricing {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 10px;
+}
+
+.discounted-price {
+  color: #ff4d4f;
+  font-weight: bold;
+  margin-right: 10px;
+}
+
+.original-price {
+  text-decoration: line-through;
+  color: #aaa;
+}
+
+.product-rating {
+  margin-top: 10px;
+}
+
+.add-to-cart {
+  margin-top: 10px;
+  background-color: #1890ff;
+  border-color: #1890ff;
+  color: white;
+}
